web/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
paho-mqtt>=2.0.0
jinja2>=3.1.0
python-dotenv>=1.0.0
